full notification routes

+++ src/main/java/com/simple_bank_transfer/account/dto/AccountDto.java
@@ -3,6 +3,6 @@ package com.simple_bank_transfer.account.dto;
 import lombok.Builder;
 
 @Builder
-public record AccountDto(Long id, String name, Integer balance) {
+public record AccountDto(Long id, String name, Long balance) {
 }
